% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{addContext}
\alias{addContext}
\title{Find the neighboring bases of a cytosine in a Granges object}
\usage{
addContext(met_locs, genome = Mmusculus, width = 5)
}
\arguments{
\item{met_locs}{GRanges object indicating the mehtylated Cytosines}

\item{genome}{Bsgenome, Reference Genome of the alignment}
}
\value{
GRanges object with mehtylated Cytosines and added context metadata
}
\description{
Find the neighboring bases of a cytosine in a Granges object
}
\examples{
ml <- GRanges(seqnames = 'chr1',
              IRanges(start=c(3508273, 3508274, 3508275), width = 1),
              strand = c("-", "-", "+"))
addContext(ml)

}
