% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{addType}
\alias{addType}
\title{Add a type column to GRanges object of a methylation call by greping contexts}
\usage{
addType(met_locs, motifs = c(acc_motif = DNAString("DGCHN"), endmet_motif =
  DNAString("NWCGW")))
}
\arguments{
\item{met_locs}{GRanges object indicating the mehtylated Cytosines}

\item{motifs}{Vector of named DNAstrings that contain the motifs to extract}
}
\value{
GRanges object with added metadata columns indicating the type
}
\description{
Add a type column to GRanges object of a methylation call by greping contexts
}
