% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixUtils.R
\name{colMeans_drop0}
\alias{colMeans_drop0}
\title{Compute column means of sparse matrix (dgC format) of nonzero elements}
\usage{
colMeans_drop0(dgCMat)
}
\arguments{
\item{dgCMat}{matrix of class dgCMatrix}
}
\value{
vector of column means. If there is no nonzero element in a column, will return NA
}
\description{
Compute column means of sparse matrix (dgC format) of nonzero elements
}
