% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{filterByConvRate}
\alias{filterByConvRate}
\title{Filter reads using nonCpG/GpC sites as a refrerence}
\usage{
filterByConvRate(met, conv_rate_cutoff = 0.8, plot = TRUE,
  keep_noinfo = TRUE)
}
\arguments{
\item{met}{Methylation data object}

\item{conv_rate_cutoff}{the minimum conversion rate per read to be kept}

\item{plot}{boolean, whether to plot the distribution of methylation for CpG,GpC and nonCpGGpC sites}

\item{keep_noinfo}{boolean, whether to keep reads that have no non CpG/GpC sites.
(For those reads the conversion rate cannot be estimated)}
}
\value{
met subset for reads with sufficient conversion rate
}
\description{
Filter reads using nonCpG/GpC sites as a refrerence
}
