% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMethylation.R
\name{getCMethMatrix}
\alias{getCMethMatrix}
\title{Call methylation events on all cytosines}
\usage{
getCMethMatrix(proj, range, sample)
}
\arguments{
\item{proj}{qProject object}

\item{range}{GRanges object with ONE!!!! range}

\item{sample}{name of the Genome alignment}
}
\value{
list with GRanges of Cytosine positions and Matrix with methylation events
\itemize{
  \item GRanges object describing the locations of called cytosines with some metadata for each site:
  \describe{
    \item{strand}{of the cytosine (strands can later be collapsed)}
    \item{met_rate}{Number of methylated sites divided by all sites that are called}
  }
  \item Matrix with the methylation calls for each read and genomic position. Uncalled=NA
}
}
\description{
Call methylation events on all cytosines
}
