% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixUtils.R
\name{rowMeans_drop0}
\alias{rowMeans_drop0}
\title{Compute row means of sparse matrix (dgC format) of nonzero elements}
\usage{
rowMeans_drop0(dgCMat)
}
\arguments{
\item{dgCMat}{matrix of class dgCMatrix}
}
\value{
vector of row means. If there is no nonzero element in a row, will return NA#'
}
\description{
Compute row means of sparse matrix (dgC format) of nonzero elements
}
